% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_dQdV_multi}
\alias{arbin_dQdV_multi}
\title{arbin_dQdV_multi}
\usage{
arbin_dQdV_multi(list, label, cycle = 1, ymin = NULL, ymax = NULL)
}
\arguments{
\item{list}{list of data generated from arbinimport script.}

\item{label}{character vector containing the legend for each Cell for the legend}

\item{cycle}{Number of the cycle of interest, default is cycle=1}

\item{ymin}{set ploting window range for the dQ/dV axis, defaults to NULL which uses the minimum value. May need to be adjusted manually to avoid outliers.}

\item{ymax}{set ploting window range for the dQ/dV axis, defaults to NULL which uses the maximum value. May need to be adjusted manually to avoid outliers.}
}
\description{
Compare differential capacity plots from different cells, same cycle. .
}
\examples{
arbin_dQdV_multi(l,title=c("Cell1", "Cell2","Cell3"),cycle=1,ymin=-0.01,ymax=0.01)
}
