% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{arbin_import}
\alias{arbin_import}
\title{arbin_import}
\usage{
arbin_import(file, step.time = TRUE, energy = TRUE, cycles = 100,
  mass = NULL, meanE = FALSE, mass.err = 0.02, electr.err = 0.025)
}
\arguments{
\item{file}{The filename, which must end in .xls or .xlsx.}

\item{step.time}{Defaults to TRUE. Includes the step time variable from the data file if TRUE.}

\item{energy}{Defaults to TRUE. Includes the (dis)charge energy variables from the data file if TRUE.}

\item{cycles}{Defaults to 100. Determines the maximum number of cycles to be considered when
aggregating the statistics dataset.}

\item{mass}{Defaults to NULL. If an active material mass is specified - in MILLIGRAMS - the
capacities in the raw and statistics data frames will be converted to mAh/g.}

\item{meanE}{Defaults to FALSE. Will calculate a statistic for average charge and discharge voltage 
(as energy per cycle divided by charge) if set to TRUE. Note: energy must also be set to TRUE - 
if it is not, the average voltage will not be calculated.}

\item{mass.err}{Defaults to 0.02 (milligrams). Is the error associated with weighing an electrode coating.}

\item{electr.err}{Defaults to 0.025 (milligrams). Is the error associated with the weight of
the electrode substrate/foil itself.}
}
\description{
This function takes an exported data file in Microsoft Excel format,
discards or includes certain data depending on the options chosen in the function,
and returns a list of two data frames: the complete raw data and an aggregated
statistics file. The statistics are aggregated from the raw data alone; the statistics
sheet as outputted to the Excel file is not read.
}
\examples{
mydataset <- arbin_import("dataset.xlsx")
mydataset <- arbin_import("dataset.xlsx", step.time = FALSE, cycles = 200, mass = 2.55)
}
\keyword{import}

