% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{arbin_import_folder}
\alias{arbin_import_folder}
\title{arbin_import_folder}
\usage{
arbin_import_folder(path = ".", mass = NULL, area = NULL, vol = NULL)
}
\arguments{
\item{path}{The folder path, which must include .xls or .xlsx files produced by the arbin.}

\item{mass}{Defaults to NULL. A vector including the active masses (in MILLIGRAMS, mg) of all cells in the folder being imported.
Use arbin_import_folder_check to ensure correct order.}

\item{area}{Defaults to NULL. A vector including the area (in SQUARE CENTIMETERS, cm^2) of all cells in the folder being imported.
Use arbin_import_folder_check to ensure correct order.}

\item{vol}{Defaults to Null. A vector including the volume (in cubic CENTIMETERS, cm^3) of all cells in the folder being imported.
Use arbin_import_folder_check to ensure correct order.}
}
\description{
This function is a simplified version of arbin_import which does not output a separate
statistics data frame. Consequently the output is a data frame rather than a list.
}
\examples{
mydataset <- arbin_import_folder(path)
mydataset <- arbin_import_folder()
}
