% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{arbin_import_folder_check}
\alias{arbin_import_folder_check}
\title{arbin_import_folder_check}
\usage{
arbin_import_folder_check(path = ".")
}
\arguments{
\item{path}{The folder path containing only data files from the arbin, which must end in .xls or .xlsx.}
}
\description{
This function is a simplified version of arbin_import which does not output a separate
statistics data frame. Consequently the output is a data frame rather than a list.
}
\examples{
mydataset <- arbin_import_folder(path)
mydataset <- arbin_import_folder()
}
