% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_plotvp}
\alias{arbin_plotvp}
\title{arbin_plotvp}
\usage{
arbin_plotvp(data, cycles = 1, norm = NULL, title = NULL)
}
\arguments{
\item{data}{The dataset, which is the list as output by arbin_import.}

\item{cycles}{The cycles to be plotted, expressed as a vector. Defaults to cycle = 1.}

\item{norm}{character string used to select how to normalize your data: "mass", "area", "vol". Deaults to Null which plots absolute capacity.}

\item{title}{Character vector to use for the title of the plot. Defaults to NULL.}
}
\description{
This function takes a data frame of raw data and a specified cycle - or number
of cycles, as a vector - and outputs charge and discharge voltage profiles.
}
\examples{
arbin_plotvp(mydataset, cycles=1)
arbin_plotvp(mydataset, cycles = c(1,5,10), norm="mass")
}
