% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_plotvp_multi}
\alias{arbin_plotvp_multi}
\title{arbin_plotvp_multi}
\usage{
arbin_plotvp_multi(list, labels, cycle = 1)
}
\arguments{
\item{list}{list containing list of data from multiple cells generated by arbinimport.}

\item{labels}{character vector containing the text to use for the plot legend. Must be as long as the number of cells being plotted}

\item{cycle}{The number of the cycle that is desired to be plotted. Default is cycle =1.}
}
\description{
Plot voltage vs capacity for multiple cells at one cycle. Used for rate capability
comparison and polarization comparison of multiple cells.
}
\examples{
arbin_plotvp_multi(l1[1:3],c("test","test2","test3"),cycle=5)
}
