% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmap_function.R
\name{arbin_process_Rmap}
\alias{arbin_process_Rmap}
\title{arbin_process_Rmap}
\usage{
arbin_process_Rmap(data.in)
}
\arguments{
\item{data.in}{The dataset to process. Can be either in list (arbin_import) or data.frame
(arbin_import_raw) form.}
}
\description{
This function takes a data set for a "resistance mapping" experimentimported with either arbin_import
or arbin_import_raw and returns a processed data containing resistance information as well as
voltage, capacity, cycle number etc.
}
\examples{
processed <- arbin_process_Rmap(rawdata)

}
\keyword{map}
\keyword{process}
\keyword{resistance}

