% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_quickplot}
\alias{arbin_quickplot}
\title{arbin_quickplot}
\usage{
arbin_quickplot(data, x, y, geom = geom_point, size = 4)
}
\arguments{
\item{data}{The dataset, which must be a dataframe.}

\item{x}{The variable to be plotted on the x-axis}

\item{y}{The variable to be plotted on the y-axis}

\item{geom}{The geom to be passed to ggplot; e.g. geom_point or geom_path}

\item{size}{The size of the geom.}
}
\description{
This function takes a data frame of either raw data or aggregated statistics,
any argument for x or y-axis plotting, and returns a ggplot with the desired
formatting and axis labels. Can also do things such as subsetting data with
dplyr.
}
\examples{
arbin_quickplot(mydataset$raw, x = t, y = E)
arbin_quickplot(filter(mydataset$raw, cyc.n == 1, x = t, y = E, geom = geom_path, size = 1)
arbin_quickplot(mydataset$stat, x = cyc.n, y = d.Q)
}

