% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmap_function.R
\name{arbin_Rmap}
\alias{arbin_Rmap}
\title{arbin_Rmap}
\usage{
arbin_Rmap(data.in)
}
\arguments{
\item{data.in}{A data.frame or list imported with arbin_import or arbin_import_raw respectively.}
}
\description{
This function is for processing "galvanostatic cycling with current interruption" (GCCI, aka "resistance
mapping") data which has been imported with either of the arbin_import functions. This function is designed to work
with only very specific Arbin schedules. This function takes only an imported data file (as a data.frame or list of raw and
statistics) and returns a processed data file with calculated internal resistances and errors, among other required data.
}
\examples{
mydataset <- arbin_import("dataset.xlsx", mass = 3)
R_data <- arbin_Rmap(mydataset)
}
\keyword{Rmap}

