% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assignRest}
\alias{assignRest}
\title{assignRest}
\usage{
assignRest(x)
}
\arguments{
\item{x}{The vector outputted by signCurrent to be processed.}
}
\description{
Takes a vector of character values outputted by the signCurrent function and assigns a unique
identifier for each distinct current-rest cycle. Called as part of the arbin_process_Rmap function.
}
\examples{
raw$rests <- assignRest(raw$state)
}

