% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmp_functions.R
\name{eclab_import_gcpl}
\alias{eclab_import_gcpl}
\title{eclab_import_gcpl}
\usage{
eclab_import_gcpl(file, mass = NULL, cycles = 100)
}
\arguments{
\item{file}{The filename}

\item{mass}{Defaults to NULL. If an active material mass is specified - in MILLIGRAMS - the
capacities will be converted to mAh/g.}

\item{cycles}{Defaults to 100. Determines the maximum number of cycles to be considered when
aggregating the statistics dataset.}
}
\description{
This function takes .mpt or .txt files exported from EC-Lab using the "Process Data" option. 
Cycle number must be included, else the function will stop prematurely. The function will
returns selected columns only (time, cycle number, step number, current, voltage, capacity), and
makes a separate 'stats' data frame similar to arbin_import.
}
\examples{
mydataset <- eclab_import_gcpl("dataset.mpt")
mydataset <- eclab_import_gcpl("dataset.mpt", mass = 3.5, cycles = 50)
}
\keyword{import}

