% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmp_functions.R
\name{eclab_importall}
\alias{eclab_importall}
\title{eclab_importall}
\usage{
eclab_importall(file, mass = NULL)
}
\arguments{
\item{file}{The filename}

\item{mass}{Defaults to NULL. If an active material mass is specified - in MILLIGRAMS - the
capacities will be converted to mAh/g.}
}
\description{
This function takes .mpt or .txt files exported from EC-Lab using the "Process Data" option. 
Cycle number must be included, else the function will stop prematurely. The function will
return *all* columns imported, but will rename common ones (E, I, Q.d etc) so as to be
consistent with the arbin_import functions. The function also checks whether Q is in coulombs,
and converts to mAh if so.
}
\examples{
mydataset <- eclab_importall("dataset.mpt", mass = 3.5)
}
\keyword{import}

