% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmp_functions.R
\name{eclab_plotvp}
\alias{eclab_plotvp}
\title{eclab_plotvp}
\usage{
eclab_plotvp(data, cycles)
}
\arguments{
\item{data}{dataset, which should be imported with eclab_import_gcpl.}

\item{cycles.}{The cycles to be plotted, expressed as a vector}
}
\description{
Voltage profile plotting function for data imported with the eclab_import_gcpl function.
Fixes unwanted data points differently to the arbin_plotvp function and is purely experimental
at this point.
}
\examples{
eclab_plotvp(mydataset, 1)
eclab_plotvp(mydataset, cycles = c(1,5,10))
}

