% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{multiplot}
\alias{multiplot}
\title{Combine plots from many arbin plots into one for ease of report creation}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of plot variables saved from an arbin plot function}

\item{labels}{character vector with the legends to use in the plots}
}
\description{
Combine plots from many arbin plots into one for ease of report creation
}
\examples{
p1<-arbin_Qplot(l1[1:3],c("Cell1", "Cell2","Cell3"))
p2<-arbin_Qplot(l1[4:6],c("Cell4", "Cell5","Cell6"))
p3<-arbin_quickplot(l1[[1]],x="Q.d",y="E",norm="mass"))
multiplot(p1,p2,p3,cols=1)
multiplot(p1,p2,p3,cols=2)
}
