% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{signCurrent}
\alias{signCurrent}
\title{signCurrent}
\usage{
signCurrent(x)
}
\arguments{
\item{x}{The vector of current values to be processed}
}
\description{
Takes a vector of current values from a data frame of raw data, and assigns a letter
("D", for discharging, "C", for charging or "R" for resting) based on the sign of the current.
This function is called as part of the arbin_process_Rmap function.
}
\examples{
raw$state <- signCurrent(raw$I)
}

