% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{theme_arbintools}
\alias{theme_arbintools}
\title{theme_arbintools}
\usage{
theme_arbintools(base_size = 16)
}
\description{
This is a ggplot theme for the arbintools package. There are no arguments.
}
\examples{
p <- p + theme_arbintools()
}

