% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_dQdV}
\alias{arbin_dQdV}
\title{arbin_dQdV}
\usage{
arbin_dQdV(list, title = NULL, cycle = 1, ymin = -0.1, ymax = 1)
}
\arguments{
\item{list}{list of data generated from arbinimport script.}

\item{title}{input legend for the plot}

\item{cycle}{number of the cycle of interest, defaults to 1}

\item{ymin}{set ploting window range, defaults to 0.1}

\item{ymax}{set ploting window range, defaults to 1}
}
\description{
Plot differential capacity plots (dQ/dV) for one cell at one cycle.
}
\examples{
arbin_dQdV(list,cellfile,1,ymin,ymax)
}
