% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_dQdV_multi}
\alias{arbin_dQdV_multi}
\title{arbin_dQdV_multi}
\usage{
arbin_dQdV_multi(list, title, cycle = 1, ymin = -1000, ymax = 1500)
}
\arguments{
\item{list}{list of data generated from arbinimport script.}

\item{title}{character vector containing the legend for each Cell for the legend}

\item{cycle}{Number of the cycle of interest, default is cycle=1}

\item{ymin}{set ploting window range for the dQ/dV axis, default is -1000}

\item{ymax}{set ploting window range for the dQ/dV axis, default is 1500}
}
\description{
Compare differential capacity plots from different cells, same cycle. .
}
\examples{
arbin_dQdV_multi(l,title=c("Cell1", "Cell2","Cell3"),cycle=1,ymin=-1000,ymax=1500)
}
