% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{arbin_import_raw}
\alias{arbin_import_raw}
\title{arbin_import_raw}
\usage{
arbin_import_raw(file, step.time = TRUE, energy = TRUE, mass = NULL)
}
\arguments{
\item{file}{The filename, which must end in .xls or .xlsx.}

\item{step.time}{Defaults to TRUE. Includes the step time variable from the data file if TRUE.}

\item{energy}{Defaults to TRUE. Includes the (dis)charge energy variables from the data file if TRUE.}

\item{mass}{Defaults to NULL. If an active material mass is specified - in MILLIGRAMS - the
capacities in the raw and statistics data frames will be converted to mAh/g.}
}
\description{
This function is a simplified version of arbin_import which does not output a separate
statistics data frame. Consequently the output is a data frame rather than a list.
}
\examples{
mydataset <- arbin_import("dataset.xlsx")
mydataset <- arbin_import("dataset.xlsx", step.time = FALSE)
}
