% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_plotvp}
\alias{arbin_plotvp}
\title{arbin_plotvp}
\usage{
arbin_plotvp(data, cycles)
}
\arguments{
\item{data}{The dataset, which can be the list as outputted by arbin_import or the data
frame as arbin_import_raw.}

\item{cycles}{The cycles to be plotted, expressed as a vector}
}
\description{
This function takes a data frame of raw data and a specified cycle - or number
of cycles, as a vector - and outputs charge and discharge voltage profiles.
}
\examples{
arbin_plotvp(mydataset, 1)
arbin_plotvp(mydataset, cycles = c(1,5,10))
}
