% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{arbin_quickplot}
\alias{arbin_quickplot}
\title{arbin_quickplot}
\usage{
arbin_quickplot(l, x = "cyc.n", y = "Q.d", norm = NULL,
  geom = geom_point, size = 4)
}
\arguments{
\item{x}{The variable to be plotted on the x-axis}

\item{y}{The variable to be plotted on the y-axis}

\item{norm}{character string used to select how to normalize your data: "mass", "area", "vol".}

\item{geom}{The geom to be passed to ggplot; e.g. geom_point or geom_path}

\item{size}{The size of the geom.}

\item{data}{The dataset, which must be a list generated by arbinimport.}
}
\description{
This function takes a data frame of either raw data or aggregated statistics,
any argument for x or y-axis plotting, and returns a ggplot with the desired
formatting and axis labels. Can also do things such as subsetting data with
dplyr.
}
\examples{
arbin_quickplot(list, x = t, y = E)
arbin_quickplot(list, x = t, y = E, norm="mass", geom = geom_path, size = 1)
arbin_quickplot(list, x = cyc.n, y = d.Q, norm="area")
}
